No motor functions, fix and praise the sun

+++ Node_2/lib/PID/PID.h
@@ -23,7 +23,7 @@
 #define F_CPU 16000000
 
 void PID_init(void);
-uint8_t PID_scale(int8_t encoder_val);
+uint16_t PID_scale(int16_t encoder_val);
 int16_t PID_control(can_message_t msg);
 uint16_t Slider_to_position_scale(int8_t encoder_val);
 void PID_update_ref(int16_t pos);
